requirements.txt
PyYAML~=6.0.1
openai~=1.10.0
torch~=2.0.1
transformers~=4.36.0
chainlit~=1.0.200